[#9] chore: shell focus control 로직 훅으로 분리
